organize marionette components using browserify
