+++ app/js/index.js
@@ -1,58 +1,17 @@
-var Marionette = require('backbone.marionette')
-    Backbone   = require('backbone')
-  , ipc        = require('ipc')
-  , _          = require('underscore')
+var ipc = electronRequire('ipc')
+
+  , Tools          = require('./collections/tools')
+  , MainLayoutView = require('./views/main_layout')
+  , ToolsView      = require('./views/tools')
   ;
 
-var Tool = Backbone.Model.extend({});
-
-var ToolsCollection = Backbone.Collection.extend({
-  url: 'https://raw.githubusercontent.com/flatiron-labs/environmentalizer_app/master/script/mac/environmentalizer.json',
-  model: Tool
-});
-
-var MainLayoutView = Marionette.LayoutView.extend({
-  template: function() {
-    return _.template("<div id='tools'></div>")({});
-  },
-
-  regions: {
-    tools: '#tools'
-  }
-});
-
-var ToolView = Backbone.Marionette.ItemView.extend({
-  template: function() {
-    return _.template("<img class='img-circle media-object pull-left' src='img/test_avatar.png' width='32' height='32'><div class='media-body'><strong>List item title</strong><p>Lorem ipsum dolor sit amet.</p></div>")({});
-  },
-
-  tagName: 'li',
-
-  className: 'list-group-item'
-});
-
-var ToolsCompositeView = Marionette.CompositeView.extend({
-  template: function() {
-    return _.template("<div class='list-group-header tools-header'><div class='header-title'><h1 class='heading'>Environmentalizer</h1></div><div class='header-button'><button class='btn btn-large btn-positive'>Apply</button></div></div><ul class='list-group'></ul>")({});
-  },
-
-  className: 'environmentalizer-tools',
-
-  childView: ToolView,
-
-  childViewContainer: 'ul.list-group',
-
-  onBeforeShow: function() {
-    this.collection.fetch();
-  }
-});
 
 var mainLayout = new MainLayoutView({
   el: '#main'
 });
 
-var toolsView = new ToolsCompositeView({
-  collection: new ToolsCollection
+var toolsView = new ToolsView({
+  collection: new Tools
 });
 
 mainLayout.render();
