+++ app/js/models/tool.js
@@ -1,0 +1,6 @@
+var Backbone   = require('backbone')
+  ;
+
+var Tool = Backbone.Model.extend({});
+
+module.exports = Tool;
